requirement.txt
Package                      Version
---------------------------- ---------
accelerate                   1.7.0
annotated-types              0.7.0
beautifulsoup4               4.13.4
cachetools                   5.5.2
certifi                      2025.4.26
charset-normalizer           3.4.2
colorama                     0.4.6
coverage                     7.8.2
faiss-cpu                    1.11.0
filelock                     3.18.0
fsspec                       2025.3.2
google-ai-generativelanguage 0.6.15
google-api-core              2.24.2
google-api-python-client     2.169.0
google-auth                  2.40.1
google-auth-httplib2         0.2.0
google-cloud-vision          3.10.1
google-generativeai          0.8.5
googleapis-common-protos     1.70.0
grpcio                       1.72.0rc1
grpcio-status                1.71.0
httplib2                     0.22.0
huggingface-hub              0.31.2
idna                         3.10
iniconfig                    2.1.0
Jinja2                       3.1.6
joblib                       1.5.0
MarkupSafe                   3.0.2
mpmath                       1.3.0
networkx                     3.4.2
numpy                        2.2.5
packaging                    25.0
pillow                       11.2.1
pip                          25.1.1
pluggy                       1.6.0
proto-plus                   1.26.1
protobuf                     5.29.5
psutil                       7.0.0
pyasn1                       0.6.1
pyasn1_modules               0.4.2
pydantic                     2.11.5
pydantic_core                2.33.2
pyparsing                    3.2.3
pytesseract                  0.3.13
pytest                       8.3.5
python-dotenv                1.1.0
PyYAML                       6.0.2
regex                        2024.11.6
requests                     2.32.3
rsa                          4.9.1
safetensors                  0.5.3
scikit-learn                 1.6.1
scipy                        1.15.3
sentence-transformers        4.1.0
setuptools                   80.4.0
soupsieve                    2.7
sympy                        1.14.0
threadpoolctl                3.6.0
tokenizers                   0.21.1
torch                        2.7.0
tqdm                         4.67.1
transformers                 4.51.3
typing_extensions            4.13.2
typing-inspection            0.4.1
uritemplate                  4.1.1
urllib3                      2.4.0